+++ src/main/java/me/lambdaurora/lambdynlights/api/DynamicLightHandler.java
@@ -16,12 +16,14 @@ import net.minecraft.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.function.Function;
+
 /**
  * Represents a dynamic light handler.
  *
  * @param <T> The type of the light source.
  * @author LambdAurora
- * @version 1.2.1
+ * @version 1.2.3
  * @since 1.1.0
  */
 public interface DynamicLightHandler<T>
@@ -33,6 +35,43 @@ public interface DynamicLightHandler<T>
      * @return The luminance.
      */
     int getLuminance(T lightSource);
+
+    /**
+     * Returns whether the light source is water-sensitive or not.
+     *
+     * @param lightSource The light source.
+     * @return True if the light source is water-sensitive, else false.
+     */
+    default boolean isWaterSensitive(T lightSource)
+    {
+        return false;
+    }
+
+    /**
+     * Returns a dynamic light handler.
+     *
+     * @param luminance      The luminance function.
+     * @param waterSensitive The water sensitive function.
+     * @param <T>            The type of the entity.
+     * @return The completed handler.
+     */
+    static <T extends LivingEntity> @NotNull DynamicLightHandler<T> makeHandler(Function<T, Integer> luminance, Function<T, Boolean> waterSensitive)
+    {
+        return new DynamicLightHandler<T>()
+        {
+            @Override
+            public int getLuminance(T lightSource)
+            {
+                return luminance.apply(lightSource);
+            }
+
+            @Override
+            public boolean isWaterSensitive(T lightSource)
+            {
+                return waterSensitive.apply(lightSource);
+            }
+        };
+    }
 
     /**
      * Returns a living entity dynamic light handler.
@@ -61,26 +100,37 @@ public interface DynamicLightHandler<T>
      */
     static <T extends CreeperEntity> @NotNull DynamicLightHandler<T> makeCreeperEntityHandler(@Nullable DynamicLightHandler<T> handler)
     {
-        return entity -> {
-            int luminance = 0;
+        return new DynamicLightHandler<T>()
+        {
+            @Override
+            public int getLuminance(T entity)
+            {
+                int luminance = 0;
 
-            if (entity.getClientFuseTime(0.0F) > 0.001D) {
-                switch (LambDynLights.get().config.getCreeperLightingMode()) {
-                    case OFF:
-                        return 0;
-                    case SIMPLE:
-                        luminance = 10;
-                        break;
-                    case FANCY:
-                        luminance = (int) (entity.getClientFuseTime(0.0F) * 10.0);
-                        break;
+                if (entity.getClientFuseTime(0.0F) > 0.001D) {
+                    switch (LambDynLights.get().config.getCreeperLightingMode()) {
+                        case OFF:
+                            return 0;
+                        case SIMPLE:
+                            luminance = 10;
+                            break;
+                        case FANCY:
+                            luminance = (int) (entity.getClientFuseTime(0.0F) * 10.0);
+                            break;
+                    }
                 }
+
+                if (handler != null)
+                    luminance = Math.max(luminance, handler.getLuminance(entity));
+
+                return luminance;
             }
 
-            if (handler != null)
-                luminance = Math.max(luminance, handler.getLuminance(entity));
-
-            return luminance;
+            @Override
+            public boolean isWaterSensitive(T lightSource)
+            {
+                return true;
+            }
         };
     }
 }