+++ src/main/java/me/lambdaurora/lambdynlights/mixin/lightsource/LivingEntityMixin.java
@@ -16,6 +16,7 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 
@@ -36,8 +37,11 @@ public abstract class LivingEntityMixin extends Entity implements DynamicLightSo
             this.lambdynlights_luminance = 15;
         } else {
             int luminance = 0;
+            BlockPos eyePos = new BlockPos(this.getX(), this.getEyeY(), this.getZ());
+            boolean submergedInFluid = !this.world.getFluidState(eyePos).isEmpty();
             for (ItemStack equipped : this.getItemsEquipped()) {
-                luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(equipped, this.isSubmergedInWater()));
+                if (!equipped.isEmpty())
+                    luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(equipped, submergedInFluid));
             }
 
             this.lambdynlights_luminance = luminance;