+++ src/main/java/me/lambdaurora/lambdynlights/mixin/lightsource/PlayerEntityMixin.java
@@ -15,6 +15,7 @@ import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -40,8 +41,11 @@ public abstract class PlayerEntityMixin extends LivingEntity implements DynamicL
             this.lambdynlights_luminance = 15;
         } else {
             int luminance = 0;
+            BlockPos eyePos = new BlockPos(this.getX(), this.getEyeY(), this.getZ());
+            boolean submergedInFluid = !this.world.getFluidState(eyePos).isEmpty();
             for (ItemStack equipped : this.getItemsEquipped()) {
-                luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(equipped, this.isSubmergedInWater()));
+                if (!equipped.isEmpty())
+                    luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(equipped, submergedInFluid));
             }
 
             this.lambdynlights_luminance = luminance;